force all plugins to be saved in a single language
Since blog posts should not be translatable, we have to force the CMS into
submission a bit here. Otherwise, the plugins are saved in the user's language,
leading to seemingly unpredictable results.

This also kind of depends on divio/django-cms#1878, which allows rendering of
a placeholder in a specific language.